+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Fixed
+
+* HTTP/2 support for tunnelled proxy cases. The `http1`/`http2` arguments are now passed on to the `httpcore` proxy pool when supported by the installed version of `httpcore`.
+
 ## 0.21.1 (16th November, 2021)
 
 ### Fixed